[IMP] sale_blanket_order: Split files
